+++ examples/basic.rs
@@ -39,17 +39,47 @@
 //! Filename: "unscratch.prg"            Blocks:   7
 //! Filename: "header change.prg"        Blocks:   5
 //! Free blocks: 403
+//!
+//! This example includes some simple error handling
 
-use rs1541::{Cbm, Rs1541Error};
+use rs1541::{Cbm, DeviceError, Rs1541Error, Xum1541Error};
+use std::process::exit;
 fn main() -> Result<(), Rs1541Error> {
     env_logger::init();
     log::info!("Started logging");
 
     // Driver automatically opens on creation and closes on drop
-    let cbm = Cbm::new()?;
+    let cbm = match Cbm::new() {
+        Err(Rs1541Error::Xum1541(error)) => match error {
+            Xum1541Error::DeviceAccess { .. } | Xum1541Error::Usb(_) => {
+                println!("Failed to connect to xum1541 device\nError: {error}");
+                exit(1);
+            }
+            _ => {
+                println!("Unexpected error from xum1541\nError: {error}");
+                exit(1);
+            }
+        },
+        Err(error) => {
+            println!("Unexpected error from rs1541\nError: {error}");
+            exit(1);
+        }
+        Ok(cbm) => cbm,
+    };
 
     // Get drive information
-    let id = cbm.identify(8)?;
+    let id = cbm.identify(8);
+    let id = match id {
+        Err(Rs1541Error::Device {
+            error: DeviceError::NoDevice,
+            ..
+        }) => {
+            println!("Device 8 not detected");
+            std::process::exit(1);
+        }
+        Err(e) => return Err(e),
+        Ok(id) => id,
+    };
     println!("Drive type at device 8: {}", id);
 
     // Check drive status