+++ src/channel.rs
@@ -1,0 +1,5 @@
+/// Dedicated load channel on disk drive
+pub const CBM_CHANNEL_LOAD: u8 = 0;
+
+/// Dedicated control/command channel on disk drive
+pub const CBM_CHANNEL_CTRL: u8 = 15;