+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.0] - 2025-01-??
 ###
+- [`Cbm::scan_bus`] and [`Cbm::scan_bus_range`]
 - Added bus scan function to examples/cli.rs
 - Added examples/statusloop.rs